Save memory with prototypes (#1)
Currently, this implementation is really inefficient and is probably leading to circular references the GC can't untangle. Each instance of a data set is being created with its own unique method objects, referencing a cache inside of a closure. With these changes, now each data source creates all methods in one prototype, and each data set is just an object with a cache property deriving from that prototype.