Avoid stack overflow when computing circular dependencies
Sometimes there can be two pathways to data, depending on if an optional "shortcut" parameter was provided to the data source. If the shortcut parameter is not provided, there may be circular dependencies. Currently, the code calculating the cost errors out with a stack overflow. This fixes things so the calculated cost for a circular dependency is "Infinity" instead.